perf: Extract shared metrics to tests/perf/metrics.py
Move recall_at_k, precision_at_k, and mrr from two_vector_bench.py into
a shared metrics module so both the two-vector and cross-encoder
benchmarks use the same implementations. Add ndcg_at_k (normalized
discounted cumulative gain with binary relevance) for the cross-encoder
benchmark.

The original module re-exports the three functions so existing callers
continue to work unchanged. test_two_vector_retrieval.py now imports
metrics directly from the shared module.

Ref: #274
